axf/requirements.txt
Django==1.11.7
django-redis==4.10.0
Pillow==6.2.0
pycryptodomex==3.7.2
PyMySQL==0.9.3
python-alipay-sdk==1.10.1
pytz==2019.2
redis==3.3.10
uWSGI==2.0.18
